+++ croud/__main__.py
@@ -23,12 +23,10 @@
 
 import argh
 import colorama
+import pkg_resources
 
 from croud import __version__
 from croud.config import Configuration
-from croud.login import login
-from croud.logout import logout
-from croud.me import me
 
 
 def main():
@@ -37,7 +35,11 @@ def main():
 
     p = argh.ArghParser(prog="CROUD")
     p.add_argument("--version", action="version", version="%(prog)s " + __version__)
-    p.add_commands([me, login, logout])
+    commands = [
+        entry_point.load()
+        for entry_point in pkg_resources.iter_entry_points("croud_commands")
+    ]
+    p.add_commands(commands)
     p.dispatch()